refactor(cldr): move week_data json structs into cldr_serde
